+++ src/CleanRegex/Internal/Prepared/Template/DelimiterAware.php
@@ -1,0 +1,12 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Prepared\Template;
+
+trait DelimiterAware
+{
+    abstract protected function delimiterAware(): string;
+
+    public function suitable(string $candidate): bool
+    {
+        return \strpos($this->delimiterAware(), $candidate) === false;
+    }
+}