+++ ChangeLog.md
@@ -7,7 +7,9 @@ Incoming
 * Bug fixes
     * Fixed a bug, where using `match()->filter()` didn't throw `InvalidReturnValueException`.
     * Fixed a bug, where using `group()->filter()` didn't throw `InvalidReturnValueException`.
-    * Fixed a bug, where `Pattern::template()->mask()` keywords weren't taken into account, when choosing a delimiter.
+    * Fixed a bug, where `Pattern::template()->mask()` keywords weren't taken into account, when choosing a delimiter
+* Features
+    * Added `Pattern::template()->pattern()`
 * Others
     * Updated `ExplicitDelimiterRequiredException` message for `Pattern::of()`
     * Updated `ExplicitDelimiterRequiredException` message for `Pattern::mask()`